land-queue/pyproject.toml
[tool.poetry]
name = "mono-repo-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "1.1.2"
PyGithub = "1.51"
jwt = "1.0.0"
requests = "^2.23.0"
ipython = "^7.14.0"
pytest = "^5.4.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
